requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==5.0.0  # https://github.com/un33k/python-slugify
Pillow==10.0.0 # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.5.0  # https://github.com/evansd/whitenoise



# Django
# ------------------------------------------------------------------------------
django==4.1.13  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.2.0 # https://github.com/adamchainz/django-cors-headers
djangorestframework-simplejwt==5.2.2 # https://github.com/jazzband/django-rest-framework-simplejwt/blob/master/docs/getting_started.rst
drf-api-logger # https://github.com/vishalanandl177/DRF-API-Logger
celery[redis]==5.3.1
django-celery-beat==2.5.0
hiredis==2.2.3
redis==4.6.0
